requirements.txt
flask==2.3.2
python-dotenv==1.0.0
firebase-admin==6.1.0
google-cloud-firestore==2.11.1
google-generativeai==0.3.0
gunicorn==20.1.0
flask-login==0.6.2
flask-wtf==1.1.1
wtforms==3.0.1
email-validator==2.0.0